requirements.txt
asgiref==3.5.2
certifi==2022.12.7
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
numpy==1.24.1
pandas==1.5.3
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
